requirements.txt
python-telegram-bot
openai
python-dotenv
tinydb
telegramify-markdown
requests
langchain
langchain-openai
langchain-community
langchain-tavily
aiosqlite
langdetect
requests
beautifulsoup4
reportlab
openpyxl
python-docx
pypdf